+++ src/Calamity/Internal/BoundedStore.hs
@@ -12,20 +12,20 @@ import           Control.Lens
 import           Control.Monad.State.Lazy
 
 import           Data.Default.Class
-import           Data.Generics.Labels                 ()
-import           Data.HashMap.Lazy                    ( HashMap )
-import qualified Data.HashMap.Lazy                    as H
+import           Data.Generics.Labels     ()
+import           Data.HashMap.Lazy        ( HashMap )
+import qualified Data.HashMap.Lazy        as H
 
-import qualified Deque.Lazy                           as DQ
-import           Deque.Lazy                           ( Deque )
+import qualified Deque.Lazy               as DQ
+import           Deque.Lazy               ( Deque )
 
 import           GHC.Generics
 
 data BoundedStore a = BoundedStore
   { itemQueue :: Deque (Snowflake a)
   , items     :: HashMap (Snowflake a) a
-  , limit        :: Int
-  , size         :: Int
+  , limit     :: Int
+  , size      :: Int
   }
   deriving ( Show, Generic )
 
@@ -36,6 +36,7 @@ instance Default (BoundedStore a) where
   def = BoundedStore mempty H.empty 1000 0
 
 type instance (Index (BoundedStore a)) = Snowflake a
+
 type instance (IxValue (BoundedStore a)) = a
 
 instance HasID' a => Ixed (BoundedStore a)
@@ -51,14 +52,14 @@ instance HasID' a => At (BoundedStore a) where
 
 addItem :: HasID' a => a -> BoundedStore a -> BoundedStore a
 addItem m = execState $ do
-  unlessM (H.member (getID m) <$> use #items) do
+  unlessM (H.member (getID m) <$> use #items) $ do
     #itemQueue %= DQ.cons (getID m)
     #size += 1
 
   size <- use #size
   limit <- use #limit
 
-  when (size > limit) do
+  when (size > limit) $ do
     q <- use #itemQueue
     let Just (rid, q') = DQ.unsnoc q
     #itemQueue .= q'
@@ -76,7 +77,7 @@ getItem id s = H.lookup id (s ^. #items)
 
 dropItem :: Snowflake a -> BoundedStore a -> BoundedStore a
 dropItem id = execState $ do
-  whenM (H.member id <$> use #items) do
+  whenM (H.member id <$> use #items) $ do
     #size -= 1
 
   #itemQueue %= DQ.filter (/= id)